# springdoc-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/user/create": {
      "post": {
        "tags": [
          "User Controller"
        ],
        "summary": "Create a new user",
        "description": "Creates a new user with the provided details.",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperUserResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/status/standard/create": {
      "post": {
        "tags": [
          "Status Controller"
        ],
        "summary": "Create a standard status",
        "description": "Creates a new standard status record in the system with the provided details.",
        "operationId": "createStandardStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStandardStatusRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardStatusResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/status/document/create": {
      "post": {
        "tags": [
          "Status Controller"
        ],
        "summary": "Create a document status",
        "description": "Creates a new document status entry for tracking document workflow states.",
        "operationId": "createDocumentStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentStatusRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperDocumentStatusResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standard/create": {
      "post": {
        "tags": [
          "Standards"
        ],
        "summary": "Create a new standard",
        "description": "Creates a new ISO standard using the provided details.",
        "operationId": "createStandard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStandardRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/update": {
      "patch": {
        "tags": [
          "User Controller"
        ],
        "summary": "Update an existing user",
        "description": "Applies partial updates to an existing user based on the provided data.",
        "operationId": "applyGeneralUpdates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperUserResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customer/update": {
      "patch": {
        "tags": [
          "Customers"
        ],
        "summary": "Update an existing customer",
        "description": "Updates one or more fields of an existing customer record.",
        "operationId": "applyGeneralUpdates_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperCustomerResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/all": {
      "get": {
        "tags": [
          "User Controller"
        ],
        "summary": "Fetch all users with pagination",
        "description": "Retrieves a paginated list of users. Default page=0, size=10.",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperListUserResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standard/all": {
      "get": {
        "tags": [
          "Standards"
        ],
        "summary": "Get all standards with pagination",
        "description": "Retrieves a paginated list of all ISO standards. Default page=0, size=20.",
        "operationId": "getAllStandards",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperListStandardResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/delete/{id}": {
      "delete": {
        "tags": [
          "User Controller"
        ],
        "summary": "Delete a user",
        "description": "Deletes an existing user from the system by their ID.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateUserRequestDTO": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 30,
            "minLength": 10,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseWrapperUserResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserResponseDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateStandardStatusRequestDTO": {
        "required": [
          "statusName"
        ],
        "type": "object",
        "properties": {
          "statusName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ApiResponseWrapperStandardStatusResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StandardStatusResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StandardStatusResponseDTO": {
        "type": "object",
        "properties": {
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "statusName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateDocumentStatusRequestDTO": {
        "required": [
          "statusName"
        ],
        "type": "object",
        "properties": {
          "statusName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ApiResponseWrapperDocumentStatusResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentStatusResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocumentStatusResponseDTO": {
        "type": "object",
        "properties": {
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "statusName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateStandardRequestDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "publishedDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ApiResponseWrapperStandardResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StandardResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StandardResponseDTO": {
        "type": "object",
        "properties": {
          "standardId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "publishedDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseWrapperCustomerResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CustomerResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CustomerResponseDTO": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PatchFieldBoolean": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "PatchFieldInteger": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "PatchFieldString": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "UpdateUserRequestDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/PatchFieldInteger"
          },
          "firstName": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "lastName": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "isActive": {
            "$ref": "#/components/schemas/PatchFieldBoolean"
          }
        }
      },
...
      "UpdateCustomerRequestDTO": {
        "type": "object",
        "properties": {
          "customerId": {
            "$ref": "#/components/schemas/PatchFieldInteger"
          },
          "name": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "email": {
            "$ref": "#/components/schemas/PatchFieldString"
          }
        }
      },
      "ApiResponseWrapperListUserResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDTO"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseWrapperListStandardResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StandardResponseDTO"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseWrapperVoid": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}